# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AccessibilityResult {
  success: boolean;
  sessionId: string;
  transcript: string;
  captions: string;
  audioDescription: string;
}

export interface ErrorResponse {
  success: boolean;
  error: string;
}

export type GenerateAccessibilityBody = {
  /** Video file to process (mp4, mov, webm, m4v) */
  video: Blob;
};

export type DownloadFileParams = {
  /**
   * Session ID from generate response
   */
  sessionId: string;
};